+++ horde/enums.py
@@ -29,4 +29,5 @@ class UserRoleTypes(enum.Enum):
     CUSTOMIZER = 4
     VPN = 5
     SPECIAL = 6
+    SERVICE = 7