+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# 4.25.0
+
+* Added service account role
+* service account can now report their proxied accounts.
+* Validates IP addresses added to the bans
+* Added min_p to LLM payloads
+
 # 4.24.0
 
 * Notifies mods on too many gens being censored as csam